#9 - Fix prettier and separate into files
